// gauge-ts/src/gen/spec.ts
export interface Span {
  start: string;
  end: string;
  startChar: string;
  endChar: string;
}
// ...
export interface ProtoStepValue {
  /** / The actual string value describing he Step */
  stepValue: string;
  /** / The parameterized string value describing he Step. The parameters are replaced with placeholders. */
  parameterizedStepValue: string;
  /** / A collection of strings representing the parameters. */
  parameters: string[];
}
// ...
export const Span: MessageFns<Span> = {
  encode(message: Span, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.start !== "0") {
      writer.uint32(8).int64(message.start);
    }
    if (message.end !== "0") {
      writer.uint32(16).int64(message.end);
    }
    if (message.startChar !== "0") {
      writer.uint32(24).int64(message.startChar);
    }
    if (message.endChar !== "0") {
      writer.uint32(32).int64(message.endChar);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Span {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSpan();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.start = reader.int64().toString();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.end = reader.int64().toString();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.startChar = reader.int64().toString();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.endChar = reader.int64().toString();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Span {
    return {
      start: isSet(object.start) ? globalThis.String(object.start) : "0",
      end: isSet(object.end) ? globalThis.String(object.end) : "0",
      startChar: isSet(object.startChar) ? globalThis.String(object.startChar) : "0",
      endChar: isSet(object.endChar) ? globalThis.String(object.endChar) : "0",
    };
  },

  toJSON(message: Span): unknown {
    const obj: any = {};
    if (message.start !== "0") {
      obj.start = message.start;
    }
    if (message.end !== "0") {
      obj.end = message.end;
    }
    if (message.startChar !== "0") {
      obj.startChar = message.startChar;
    }
    if (message.endChar !== "0") {
      obj.endChar = message.endChar;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Span>, I>>(base?: I): Span {
    return Span.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Span>, I>>(object: I): Span {
    const message = createBaseSpan();
    message.start = object.start ?? "0";
    message.end = object.end ?? "0";
    message.startChar = object.startChar ?? "0";
    message.endChar = object.endChar ?? "0";
    return message;
  },
};
// ...
export const ProtoStepValue: MessageFns<ProtoStepValue> = {
  encode(message: ProtoStepValue, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.stepValue !== "") {
      writer.uint32(10).string(message.stepValue);
    }
    if (message.parameterizedStepValue !== "") {
      writer.uint32(18).string(message.parameterizedStepValue);
    }
    for (const v of message.parameters) {
      writer.uint32(26).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ProtoStepValue {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseProtoStepValue();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.stepValue = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.parameterizedStepValue = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.parameters.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ProtoStepValue {
    return {
      stepValue: isSet(object.stepValue) ? globalThis.String(object.stepValue) : "",
      parameterizedStepValue: isSet(object.parameterizedStepValue)
        ? globalThis.String(object.parameterizedStepValue)
        : "",
      parameters: globalThis.Array.isArray(object?.parameters)
        ? object.parameters.map((e: any) => globalThis.String(e))
        : [],
    };
  },

  toJSON(message: ProtoStepValue): unknown {
    const obj: any = {};
    if (message.stepValue !== "") {
      obj.stepValue = message.stepValue;
    }
    if (message.parameterizedStepValue !== "") {
      obj.parameterizedStepValue = message.parameterizedStepValue;
    }
    if (message.parameters?.length) {
      obj.parameters = message.parameters;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ProtoStepValue>, I>>(base?: I): ProtoStepValue {
    return ProtoStepValue.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ProtoStepValue>, I>>(object: I): ProtoStepValue {
    const message = createBaseProtoStepValue();
    message.stepValue = object.stepValue ?? "";
    message.parameterizedStepValue = object.parameterizedStepValue ?? "";
    message.parameters = object.parameters?.map((e) => e) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}